+++ src/oliview/vector2.h
@@ -17,8 +17,8 @@ namespace oliview {
         float y() const;
         void set_y(float value);
 
-        float get(int index) const;
-        void set(int index, float value);
+        float get(size_t index) const;
+        void set(size_t index, float value);
 
         const float * elements() const;
         float * elements();
